% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaiduYun.manue.R
\name{BaiduYun.manue}
\alias{BaiduYun.manue}
\title{Baidu Yun Manue}
\usage{
BaiduYun.manue(CatchFile)
}
\arguments{
\item{CatchFile}{path of BaiduYunCacheFileV0.db}
}
\value{
dataframe
}
\description{
Baidu Yun Manue
}
\examples{
BaiduYun.manue()
}
