% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdy.download.R
\name{bdy.download}
\alias{bdy.download}
\title{Download}
\usage{
bdy.download(..., dir, overwrite = TRUE)
}
\arguments{
\item{...}{one or more files or directory}

\item{dir}{directory which you want save files to}

\item{overwrite}{wether to overwrite an existing file}
}
\description{
Download
}
