% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdy.upload.R
\name{bdy.upload}
\alias{bdy.upload}
\title{Download}
\usage{
bdy.upload(..., todir)
}
\arguments{
\item{...}{one or more local files or directory}

\item{todir}{directory of Baidu Yun, if missing, current wd will be treated as}
}
\description{
Download
}
\examples{
# bdy.upload('E:/baiduyupcs/BaiduPCS-Go_export_1613888309.txt',
#            'E:/baiduyupcs/BaiduPCS-Go_export_1613888451.txt')
}
